requirement.txt
asgiref==3.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-crontab==0.7.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.1
#psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
